Add K-5 Dashboard card component
This change adds a new InstUI-based card component to be used instead
of the existing DashboardCard when the canvas_for_elementary flag is
enabled. It strips out a lot of the existing functionality from the
existing DashboardCard-- some may be added back in or replaced later,
but this provides a clean slate to start from.

Also adds some plumbing to the Dashboard card drag-drop box container
to allow a card implementation to be passed-in-- this ensures that we
only have to load one component or the other without having to add
extra lazy loading steps.

Finally, this also ensures that the canvas_for_elementary flag is
populated in the correct place in js_env, and adds a flag for whether
or not a course is a homeroom to the relevant APIs. This allows us to
filter out homeroom courses from the dashboard card view, since they
will be represented differently later.

Finally finally-- also adds a link to the latest announcement on each
card if the course has announcements.

closes LS-1752
closes LS-1753
flag = canvas_for_elementary

Test plan:
  General regression of Dashboard cards:
    - With the flag off, make sure cards still look and behave as
      expected.
    - With the flag on, make sure cards can stil be re-ordered and are
      layed out similarly to the old cards.
    - Set an image for a course in course settings (turning on
      course_card_images flag if necessary), verify that it shows up
      on the dashboard card
    - Enroll a user in an unpublished course and turn on
      unpublished_courses flag if necessary. Verify that two sections
      appear on the dashboard for published/unpublished courses.
  New functionality (with flag on):
    - Create a course and check the "homeroom" flag in course settings
      (or set it on an existing course)
    - Expect that course to no longer show up on the K-5 Dashboard
      homeroom tab
    - Add an announcement to a non-homeroom course (or ensure it
      already has one), verify that the announcement appears on the
      associated card
    - Click on the announcement link and ensure that it takes you to
      that specific announcement on the course page
    - Find a non-homeroom course with no announcements or create a new
      one
    - Verify that no announcement link shows up on the course card

Change-Id: Ia09110b86fc4c12e9d9a9c491ba4f3b1a4696044
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/257902
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Ed Schiebel <eschiebel@instructure.com>
QA-Review: Ed Schiebel <eschiebel@instructure.com>
Product-Review: Peyton Craighill <pcraighill@instructure.com>