Remove Intl polyfills (no longer needed on Node 14)
Closes FOO-1582
flag=none

When we were on Node 10 and 12, ICU was not supported, even
though all our target browsers do. So testing things that made
references to the formatters in window.Intl required polyfills
so the test runners would work.

Now that we're on 14, that's not needed any more, so let's
rip them out for the best test fidelity.

I left a skeleton in place for IntlPolyfills.js in case we run
into anything in the future that Node's ICU does not support
but our browsers do. Hopefully this will never happen.

Test plan:
* The "countdown" in QRMobileLogin still works
* All JS tests still pass

Change-Id: Ib55ab2bdfc5376413d339c52ffd314cb4ee3ee64
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/258079
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Ahmad Amireh <ahmad@instructure.com>
Product-Review: Ahmad Amireh <ahmad@instructure.com>
Reviewed-by: Ahmad Amireh <ahmad@instructure.com>