Verify event.focusedEditor is present for activeClass
The handleBlur event occasionally has an undefined
focusedEditor but the focus is still on the toolbar
button. This extends that check to look at both that
case and the original check if the target and focusedEditor
are the same.

fixes LS-1804
flag=none

test plan:
- In outcomes, create a new outcome
- In the new outcome form, open and close the RCE content tray
- It should always open and close and not close itself
  without input from the user

Change-Id: I138821c12f8312190e4a4ffcc4fa0af3c6812a59
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/258343
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Robin Kuss <rkuss@instructure.com>
Reviewed-by: Jonathan Guardado <jonathan.guardado@instructure.com>
QA-Review: Robin Kuss <rkuss@instructure.com>
Product-Review: Eric Saupe <eric.saupe@instructure.com>