Fix bug: setting AGS scores back-scales scores
In lti/ims/scores_controller.rb#update_or_create_result, we score the
submission with the score scaled to the assignment's points possible,
and then update the Lti::Result's score. e.g., if the assignment's
points possible is 2, and we score with result_score=9 and
result_maximum=10, the submission's will be set to 1.8 (this is what
should be shown in speedgrader), and then the result's result_score set
to 9. The problem is updating submission triggers a hook which also
updates the result's score, using the wrong (scaled) score, (e.g. it
sets result_score=1.8). This code path is only meant for setting the
result's score when the submission is manually scored.

The problem is usually corrected when ScoresController re-sets the
result's score using the unscaled score (e.g. 9), but this doesn't
happen in scenarios where the unscaled score isn't being modified (e.g.
only the maximum is changing), because the model object hasn't been
reloaded since the submission update changed the database record. That
is, `result.update!` on scores_controller.rb:329 doesn't set
result_score to 9 because it thinks result.result_score is already 9.

refs INTEROP-6399
flag=none

Test plan:
- Run these steps both before and after this change to see the bug in
  action:
- Create an assignment with AGS with non-zero points_possible (e.g. 2)
- Set a score with AGS (e.g. score=9, max=9)
- Set a score again with AGS, this time with the same score but
  different max (e.g. score=9, max=10). This means the submission's
  score (which is scaled) should change but the result's score should
  not.
- Retrieve the score from AGS:
  - Under the old code, it will show the score incorrectly scaled to the
    points_possible (e.g., score=1.8 and max=10). This it because the
    submission update triggers a SQL update to change the line result's
    score.
  - Under the new code, it will show the score correctly (score=9 and
    max=10)

Change-Id: I9a1fb790e1639bfdca640d6ce8b9be3bbdb224fd
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/258661
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Product-Review: Evan Battaglia <ebattaglia@instructure.com>
Reviewed-by: Xander Moffatt <xmoffatt@instructure.com>
QA-Review: Xander Moffatt <xmoffatt@instructure.com>