add eportfolio account report
This adds an ePortfolio report that can be ran on a root account
or any sub-account. It defaults to giving a report of all active
eportfolios that reside under that institution. There is an
option to run the report to collect ePortfolios for users that
have no enrollments or just ePortfolios that are in a deleted
state.

closes FOO-1647
flag = none

test plan:
 • Navigate to Account Settings --> Reports
 • Verify the newly added "Eportfolio Report" appears
 • Clicking on the (?) gives further details about the report
 • There are two configuration options for the report
    - Only include ePortfolios from users with no enrollments
    - Only include ePortfolios that have been removed
 • The default report with no options selected will generate
   all active ePortfolios for the given institution
 • The first config option should only include ePortfolios from
   users with no enrollments
 • The second option should only include deleted ePortfolios
 • Using both options will generate a report of deleted
   ePortfolios from users with no enrollments

Change-Id: I9d8f4cdbe354a483380d8b80256cfb976fa5a953
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/259515
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Charley Kline <ckline@instructure.com>
Product-Review: Charley Kline <ckline@instructure.com>
QA-Review: Charley Kline <ckline@instructure.com>