LTI icon shows as a 'bomb' instead of the tool's icon
When adding the icons for the LTIs in the RCE, there is a parse
error trying to assign an object when a string is expected, the
provided fix is to validate the parameter before the assignation.
For fixing the second problem about blocking the possibility to add
an LTI with ‘Editor button’ placement inactive to the RCE, some
callbacks and validations were added when the placement is set
to active/inactive.

fixes LS-1940
flag=none
 
Test Plan: 
Issue 1:

- Having a course with RCE Enhancements enabled
- Go to your account configuration and add one LTI to RCE
- Set at least one of the placements for the added LTI to inactive
- Create or edit a course to display the RCE
- Then the LTI icon should be rendered in the toolbar

Issue 2:

- Go to your account configuration and add another LTI to the RCE
- Set the ‘Editor Button’ placement for the added to inactive
- Then the toggle button is removed from the UI
- Change the ‘Editor Button’ placement to active
- Then the toggle button is added back to the UI
- Add a LTI to RCE and then set the ‘Editor Button’ placement to
inactive
- Then the LTI will be removed from the favorite apps list and
the toggle button will be removed from the UI

Change-Id: I83599f6a6d3feff2ef403b9fcd18e7d0b5626a51
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/260197
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Nate Armstrong <narmstrong@instructure.com>
QA-Review: Nate Armstrong <narmstrong@instructure.com>
Product-Review: Jonathan Guardado <jonathan.guardado@instructure.com>