Limit rendering To Do tab to supported contexts
PlannerNotes can be added to a student or to a course within
the context of a student. Groups are not supported but were
rendering the To Do tab and the form with the group if that
was all that was selected. This changes the logic for rendering
the To Do tab and the select options to the same and removes
groups from being included at any point.

fixes LS-1855
flag=none

test plan:
 - Create and publish a course
 - Enroll a student into the course
 - Create a group in the course and add the student to the group
 - As the student, go to your calendar and only have the group selected
 - In the calendar, select a day and verify the "My To Do" tab is not
  rendered
 - Select all the available calendars
 - Select a day and verify the "My To Do" tab is rendered
 - Click on the "My To Do" tab
 - Verify no student groups are available in the Calendar select

Change-Id: I6412a856bacd3cb4dfb75dc951a23900fb66ad15
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/260261
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Jeff Largent <jeff.largent@instructure.com>
Reviewed-by: Jeremy Stanley <jeremy@instructure.com>
QA-Review: Jeff Largent <jeff.largent@instructure.com>
Product-Review: Eric Saupe <eric.saupe@instructure.com>