Scope rceautosave to current user
Previously, the rceautosave local storage key was limited only to the
current window's href. This would cause different users of the same
browser to be presented with different user's autosaved data if they
went to the same content.

This does not secure this data as it is still within the browser's
local storage and could be accessed by other users. We would need to
send this data to the server for storage and retrieval to fix that.

fixes LS-1954
flag=none

test plan:
- Enable RCE Enhancements and RCE Auto Save flags
- Publish a course with teacher and students
- As a user, navigate to a page with the RCE
- Enter some data into the RCE for the autosave
- Log in as the other user
- Navigate to the same page as before
- Open the RCE and verify no prompt is shown to restore auto saved data
- Log in as the original user
- Navigate to the same page as before
- Open the RCE and verify the prompt is shown to restore auto saved data

Change-Id: I786feb3b84c4a73c39397c6b09cf96f795d1d71b
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/260408
Reviewed-by: Nate Armstrong <narmstrong@instructure.com>
QA-Review: Nate Armstrong <narmstrong@instructure.com>
Product-Review: Eric Saupe <eric.saupe@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>