Add endpoint to query visible course nav tools
This adds an endpoint that we'll use later to query which LTI tools to
show on the K5 dashboard. We want to get tools that:
 - have course_navigation placement
 - are not hidden by the teacher in navigation settings
 - have visibility settings applicable to user
The LTI tools that show up in the course navigation list satisfy all of
these requirements, so we can simply take the list of items in
course_nav (which is heavily cached), and filter for external tools.

refs LS-1770
flag=none

Test plan:
 - Hit /api/v1/courses/{id}/external_tools/visible_course_nav_tools
 - Expect to see a list of tools that meet the above requirements for
   each user doing the api call (essentially, every tool listed in
   course nav should be present in the JSON)

Change-Id: I436535e1f377b8b73c548e09acfe4c9324575910
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/260383
Reviewed-by: Jeremy Stanley <jeremy@instructure.com>
QA-Review: Jeremy Stanley <jeremy@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Product-Review: Jackson Howe <jackson.howe@instructure.com>