add initial discussion topic query
This inital query will be used by both the DiscussionContainer
and the DiscussionThreadsContainer to render real data.

Since this is just the query and we are not doing anything with
the data in this commit there isn't much to test on the frontend.
Specs are expected as we actually start using the requested data
to render the containers.

refs VICE-1238
refs VICE-1237
flag=react_discussions_post

Test Plan:
- Enable the react_discussion_post feature flag
- Open your devtools and navigate to an existing Discussion Topic
- Note the graphql network request in the networks tab

Change-Id: I30be45330058ab3d7a48520133d582e8035a9edf
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/262523
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Tested-by: Omar Soto-Fortuño <omar.soto@instructure.com>
Reviewed-by: Omar Soto-Fortuño <omar.soto@instructure.com>
QA-Review: Omar Soto-Fortuño <omar.soto@instructure.com>
Product-Review: Omar Soto-Fortuño <omar.soto@instructure.com>