Use equation image src attribute as the source of truth
closes MAT-26
flag=new_math_equation_handling

*** the same change in a sibling commit has been merged
*** into stable/2021/04/14 and is being warmfixed.

we were pulling the equation LaTex from the image's
data-equation-content attribute but in the quiz from the ticket
its value was not the equation. I'm not sure how this happened
(could the quiz author really have edited the HTMl to change
all the quiz's equations?) but since the src UTL is the equation's
source of truth, let's just always use that.

There are other places in the code that rely on the date-equation-content
attribute to generate mathml, but it's only exercised if the
new math feature is off. I'd rather not remove it from everywhere on
a ticket that will be hotfixed.

test plan:
  - with the new_math_equation_handling flag on
  - in the rce use the equation editor to create an equation
  - edit the HTML to put something else in the data-equation-content
    and alt attributes, or remove all but the src attribute if you
    want.
  - save
  > expect the correct MathJax typset equation in the result.
  - have an equation with no LaTex in the src attribute and save
  > expect the broken-image image and no catastrophic failures

Change-Id: If2a0528563804c05d8b0344cc7e4b0f1dc5a47a2
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/262593
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Nate Armstrong <narmstrong@instructure.com>
QA-Review: Nate Armstrong <narmstrong@instructure.com>
Product-Review: Ed Schiebel <eschiebel@instructure.com>