remove unwanted output from setup script in jenkins
We the addition of the spinner jenkins is showing a lot of output
coming from the spinner library. We don't need the spinner in local
dev smoker test. We add the ability to disable the spinner if is
running in jenkins.

refs DE-623
flag=none

test-plan:
1- Local dev smoker test passes and does not show unwanted output
2- setup script works locally

Change-Id: I4e476137e72ba44db69284130a1a4a7e6531dca7
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/262831
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Andrea Cirulli <andrea.cirulli@instructure.com>
Reviewed-by: James Butters <jbutters@instructure.com>
Product-Review: Andrea Cirulli <andrea.cirulli@instructure.com>