fix uniqueness constraint violation race condition
fixes FOO-1643

Pseudonym#validate_unique_id is supposed to protect against uniqueness
constraint violations, but concurrent calls for a unique_id that doesn't
yet exist can both pass the validation, which results in an unhandled
ActiveRecord::RecordNotUnique from the loser of the race to write to the
DB.

test plan:
This is hard to reproduce.  In theory, in a multi-server environment
(i.e. not local dev), you could spam the `/api/v1/accounts/:id/logins`
endpoint with parallel requests containing the same `unique_id`, and if
you're lucky, two of them might collide, giving you a 500 response on
one of them.  While that may be difficult to repro before this change,
it should be _impossible_ to repro after it.  So, run your spammer for
an infinite amount of time, and verify that it never gives you a 500.

Change-Id: If57a8a69294128a5442ad1e613b597189810e4c5
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/263213
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Jacob Burroughs <jburroughs@instructure.com>
QA-Review: Michael Ziwisky <mziwisky@instructure.com>
Product-Review: Michael Ziwisky <mziwisky@instructure.com>