Add ability to create course from k5 dashboard
Puts a button on the k5 dashboard that opens a modal to create a new
course, like on the classic dashboard. The button is visible to any
user with admin role. The modal queries a new endpoint (on open) which
returns all accounts in which the user can create courses. The user
then picks the account inside which the course is created. We also
enroll the user as a teacher in the new course to mirror what is done
in classic canvas.

This change also adds an empty state to the k5 dashboard if the user
has no enrollments - that way we can still show the 'add course'
button to admins.

closes LS-2123
flag=canvas_for_elementary

Test plan:
 - As an admin of a k5 subacount with no enrollments, visit the
   dashboard
 - Expect to see the panda
 - Click the Add Course button, expect modal to appear
 - Fill out fields and click create, expect to be redirected to the new
   course's settings page
 - Also expect to have a teacher enrollment in this new course
 - As an admin over an account with several subaccounts, visit the k5
   dashboard and open the create course modal
 - Expect to see all the possible accounts listed in the dropdown
 - Create an admin with full access to one account but remove the
   manage_courses permission on another account
 - As this admin, visit k5 dashboard and open the modal, expect to see
   the one account listed in the dropdown, but not the other

Change-Id: Idcbd997ff314a9f035862d35166505f47be9e7cd
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/263246
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Jeff Largent <jeff.largent@instructure.com>
QA-Review: Jeff Largent <jeff.largent@instructure.com>
Product-Review: Peyton Craighill <pcraighill@instructure.com>