provide frd keys to children components
refs VICE-1343
flag=none

Keys cannot be accessed on the component in which they are given. We
were attempting to do that as we had a list of elements that each were
being applied to a list of components. To fix that, we now need to
explicitly provide the key for each list item.

test plan:
  - run `yarn test:jest DiscussionThreadContainer`
    - there should be no console error regarding the `key` prop
      not being used correctly

qa risk: low

Change-Id: I7182428405ed574d6133a3d8ab79cda3992d5a79
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/263486
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Omar Soto-Fortuño <omar.soto@instructure.com>
QA-Review: Omar Soto-Fortuño <omar.soto@instructure.com>
Product-Review: Omar Soto-Fortuño <omar.soto@instructure.com>