make sure pulsar cert is on disk
and that we have a vault token

refs FOO-1842
flag=none

If a cert url is in the config, we need
to fetch and write it to disk before trying
to open a connection to pulsar or we'll
get failures.

Also add some config caching so
we aren't re-parsing the same YAML
blob over and over.

TEST PLAN:
  1) specs should pass
  2) be in a pulsar-enabled environment
      with a cert configured
  3) when you build a message_bus producer,
     the cert is fetched and written to
     disk at the configured path.

Change-Id: I2fb61badd05aedadb6d07475875692b30dd30db9
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/263687
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Jacob Burroughs <jburroughs@instructure.com>
QA-Review: Ethan Vizitei <evizitei@instructure.com>
Product-Review: Ethan Vizitei <evizitei@instructure.com>