Move a11y checker dependency into canvas-rce
closes MAT-129
flag=none

The tinymce-a11y-checker was a dependency of canvas-lms and
given to the RCE thru some fancy initialization shenanigans.
This change puts it in the new RCE, but leaves it in canvas
for the old rce. We can remove it with MAT-54 when the old
rce is removed.

test plan:
  - open a page with the rce
  - click the a11y checker button in the status bar.
  > expect the a11y checker to open.

  Make sure it still works in the old RCE
  - edit config/feature_flags/00_standard.yml and set
    rce_enhancements.environments.development.state: allowed
  - restart your canvas server
  - from /accounts/site_admin/settings page, feature options tab
    turn RCE Enhancements off
  - load a page with the rce
  > expect the a11y checker to work

Change-Id: I9d9202ee1c9a254e14bcc14298330f8a9ab8a7d4
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/264851
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Jeff Largent <jeff.largent@instructure.com>
Reviewed-by: Jeff Largent <jeff.largent@instructure.com>
Product-Review: Ed Schiebel <eschiebel@instructure.com>