don't allow siteadmin courses
Siteadmin is not a place for normal canvas things. If a course is
created on siteadmin it generally causes problems or leads to confusion
when knowing which things apply to it. This adds a validation to just
fail if the root_account is siteadmi

test plan
 - on siteadmin try to create a course
 - it should fail

flag=none
fixes VICE-1616

Change-Id: Ifd4de5403f7962cc2d79a30a4d5febd4b358c309
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/267340
Reviewed-by: Cody Cutrer <cody@instructure.com>
QA-Review: Cody Cutrer <cody@instructure.com>
Product-Review: Cody Cutrer <cody@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>