Reorganize some of the CanvasRce properties
closes MAT-328
flag=rce_enhancements

It was very confusing having 2 components named CanvasRce, so
the one in the canvas-rce package was renamed to simply RCE.

That the props for menu, toolbar, and plugins were not consistent
between canvas' CanvasRce and the RCE's CanvasRce was a source
of confusion (and just wrong). This moves them into the editorOptions
property where are all the way through the component hierarchy.

test plan:
  - from the canvas-rce directory run yarn demo:dev
  > expect the custom toolbar and menu options to work
  > expect the readonly prop to work (something was broken there too)

  - enable the site_admin feature "Assignment Enhancements - Student"
  - create a text entry assignment
  - as a student, go to the assignment and submit you work
  > expect the page to work like before

Change-Id: I982c7c54750e7a3d9e230e0ab6d16bd078ce9030
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/268505
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Nate Armstrong <narmstrong@instructure.com>
QA-Review: Nate Armstrong <narmstrong@instructure.com>
Product-Review: Ed Schiebel <eschiebel@instructure.com>